+++ CodeStarterCamp_Week1/CodeStarterCamp_step1~2.swift
@@ -1,0 +1,24 @@
+//
+//  main.swift
+//  CodeStarterCamp_Week1
+//
+//  Created by yagom.
+//  Copyright © yagom academy. All rights reserved.
+//
+
+import Foundation
+
+func icecreamHead(Head: Int) {
+    for _ in 1...Head{
+        print("***********")
+    }
+}
+
+func icecreamBody(Body: Int) {
+    for _ in 1...Body{
+        print("    | |")
+    }
+}
+
+icecreamHead(Head: 8)
+icecreamBody(Body: 4)
